[QNS] fix wrong parameter type.
Bug: 267612616
Test: atest QualifiedNetworksServiceTest

+++ services/QualifiedNetworksService/src/com/android/telephony/qns/RestrictManager.java
@@ -320,7 +320,7 @@ class RestrictManager {
 
     class LowRtpQualityRestriction extends Restriction{
         private int mReason;
-        LowRtpQualityRestriction(int type, int[] releaseEvents, int restrictTime, int reason) {
+        LowRtpQualityRestriction(int type, int[] releaseEvents, long restrictTime, int reason) {
             super(type, releaseEvents, restrictTime);
             mReason = reason;
         }
